pyproject.toml
[project]
name = "github-skills-dataset"
version = "0.1.0"
description = "Build validated SKILL.md dataset for Kaggle"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
skills-dataset = "github_skills_dataset.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_skills_dataset"]
